extract topband function
